Download suitable runtime from GitHub automatically
We no longer ship any runtimes with the AppImage, but just download the latest continuous release.

TODO: download latest *released* runtime, and allow users to specify the release they want to download a runtime for. (We need to parse the GitHub API to find the latest release; the latter variant just skips that step and replaces one component in the URL.)