Fix CA certificate chain lookup logic
When the values libcurl provides as defaults are broken, they should be
unset or replaced with working ones. Therefore, we now probe them, unset
broken ones and then run our detection logic if the values have not been
set.

This should make sure both values are either set with a seemingly usable
path or unset.

When neither of the values work, we log a warning. Then, the included
locations can be amended, if possible.