Lengthen string comparisons to unblock arm_aarch64 code path
Effectively checks that the strings being compared start and end with
the specified characters (by including the terminating NUL byte of the
fixed string), rather than only checking that a string *starts with*
the non-null ASCII characters of the fixed string in the comparison.

For example, comparing the first three characters of "pen" and "pencil"
would be a match ("pen" == "pen"). Comparing the first four bytes
instead, we get "pen[NUL byte]" != "penc".

Prevents ambiguity between "arm" and "arm_aarch64", which makes the
"arm_aarch64" code path reachable again in the large if/else of
string comparisons.